+++ src/components/common/InputField.tsx
@@ -3,8 +3,8 @@ import type { ChangeEvent, ReactNode } from 'react';
 
 import { Eye, EyeClosed } from 'lucide-react';
 
-import Input from '@/components/ui/input';
 import InputClearButton from '@/components/common/InputClearButton';
+import Input from '@/components/ui/input';
 import cn from '@/lib/utils';
 
 type InputFieldProps = React.ComponentProps<typeof Input> & {
