+++ src/components/common/SingleAvatar.tsx
@@ -1,6 +1,5 @@
+import { Avatar, AvatarImage, AvatarFallback } from '@/components/ui/avatar';
 import cn from '@/lib/utils';
-
-import { Avatar, AvatarImage, AvatarFallback } from './avatar';
 
 type SingleAvatarProps = {
   src: string;
